support float drawing

+++ mainwindow.cpp
@@ -46,7 +46,7 @@ void MainWindow::draw(QVector<Segment> segments)
     clear(ui->ruler_layout);
     QFont font_14("times",14);
     QPalette color;
-    int scale=20;
+    float scale=20;
     QLabel *zero_ruler=new QLabel("0");
     zero_ruler->setMinimumHeight(10);
     zero_ruler->setMaximumHeight(10);
@@ -60,8 +60,8 @@ void MainWindow::draw(QVector<Segment> segments)
              label=new QLabel("");
              label->setFrameStyle(1);
              color=QPalette(Qt::Window,Qt::gray);
-             label->setMinimumSize(200,scale*int(segments[i].get_to()-segments[i].get_from()));
-             label->setMaximumSize(200,scale*int(segments[i].get_to()-segments[i].get_from()));
+             label->setMinimumSize(200,scale*(segments[i].get_to()-segments[i].get_from()));
+             label->setMaximumSize(200,scale*(segments[i].get_to()-segments[i].get_from()));
              label->setAutoFillBackground(true);
              label->setPalette(color);
              label->setFont(font_14);
@@ -78,8 +78,8 @@ void MainWindow::draw(QVector<Segment> segments)
                  draw(MM.getSegments());
              });
              button->setPalette(color);
-             button->setMinimumSize(200,scale*int(segments[i].get_to()-segments[i].get_from()));
-             button->setMaximumSize(200,scale*int(segments[i].get_to()-segments[i].get_from()));
+             button->setMinimumSize(200,scale*(segments[i].get_to()-segments[i].get_from()));
+             button->setMaximumSize(200,scale*(segments[i].get_to()-segments[i].get_from()));
              ui->memory_layout->addWidget(button);
          }
          else
@@ -88,8 +88,8 @@ void MainWindow::draw(QVector<Segment> segments)
              label->setFrameStyle(1);
              QColor colour(p_color[segments[i].get_processesNames()[0]]);
              color=QPalette(Qt::Window,colour);
-             label->setMinimumSize(200,scale*int(segments[i].get_to()-segments[i].get_from()));
-             label->setMaximumSize(200,scale*int(segments[i].get_to()-segments[i].get_from()));
+             label->setMinimumSize(200,scale*(segments[i].get_to()-segments[i].get_from()));
+             label->setMaximumSize(200,scale*(segments[i].get_to()-segments[i].get_from()));
              label->setAutoFillBackground(true);
              label->setPalette(color);
              label->setFont(font_14);
@@ -102,8 +102,8 @@ void MainWindow::draw(QVector<Segment> segments)
 
          QLabel *ruler=new QLabel(QString::number(segments[i].get_to()));
          ruler->setAlignment(Qt::AlignHCenter | Qt::AlignBottom);
-         ruler->setMinimumHeight(scale*int(segments[i].get_to()-segments[i].get_from()));
-         ruler->setMaximumHeight(scale*int(segments[i].get_to()-segments[i].get_from()));
+         ruler->setMinimumHeight(scale*(segments[i].get_to()-segments[i].get_from()));
+         ruler->setMaximumHeight(scale*(segments[i].get_to()-segments[i].get_from()));
          ui->ruler_layout->addWidget(ruler);
      }
      ui->scrollArea->verticalScrollBar()->setValue(ui->scrollArea->verticalScrollBar()->minimum());
